Fix sanitizers and address violation

+++ arrays.cpp
@@ -38,9 +38,9 @@ static int InvertInplace(uint8_t *buf, size_t nbytes) {
 
   for (size_t i = 0; i < rem; i++) {
     int64_t value;
-    memcpy(&value, cursor, sizeof(int64_t));
+    memcpy(&value, cursor, sizeof(uint8_t));
     value = ~value;
-    memcpy(cursor, &value, sizeof(int64_t));
+    memcpy(cursor, &value, sizeof(uint8_t));
     cursor++;
   }
